Refactor Generator to use File instead of FileUtils.

+++ generator/Generator.h
@@ -14,19 +14,33 @@
 
 namespace lsql {
 
-	//
+	/**
+	 * Creates files with deterministic random numbers.
+	 */
 	class Generator {
 		
-		uint64_t bufferSize;
+		size_t bufferSize;
 		uint64_t seed;
 		
 	public:
 		
-		//
-		Generator(uint64_t bufferSize, uint64_t seed = 88172645463325252ull);
+		/**
+		 * Creates a new file generator.
+		 * Instead of writing to the file directly, all output is 
+		 * stored in an output buffer.
+		 *
+		 * @param bufferSize The size of the output buffer.
+		 * @param seed       A seed for the random number generator.
+		 */
+		Generator(size_t bufferSize, uint64_t seed = 88172645463325252ull);
 
-		//
-		bool generate(const char* fileName, uint64_t n) const;
+		/**
+		 * Generates a new random file at the specified location.
+		 *
+		 * @param fileName A fully qualified path to the file.
+		 * @param n        The number of elements in the file.
+		 */
+		bool generate(const char* fileName, off_t n) const;
 		
 	};
